componentized editcompleteprofile and completeprofilemodal on dashboard page
